hello_world/requirements.txt
strands-agents
awslabs.aws-documentation-mcp-server
